dataflow/requirements.txt
# DataFlow might fetch real-time market data, publish to RabbitMQ, optionally store logs in Postgres

pika==1.3.1            # RabbitMQ client library
requests==2.28.1       # For external API calls (Polygon/Alpaca data)
psycopg2-binary==2.9.7 # Postgres client, if DataFlow logs or stores metadata in DB